# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMMAS API",
        "description": "Sistem Informasi Manajemen Magang Siswa (SIMMAS). REST API dengan autentikasi Sanctum. Token didapat dari endpoint `POST /api/login`. Semua endpoint selain `login` dan `public/*` memerlukan header `Authorization: Bearer <token>`.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server backend SIMMAS (Laravel)"
        }
    ],
    "paths": {
...
        "/api/guru/siswa": {
            "get": {
                "tags": [
                    "Guru"
                ],
                "summary": "Daftar siswa bimbingan",
                "description": "Daftar penempatan siswa bimbingan guru, lengkap dengan relasi absensi, jurnal, dan DUDI tempat magang.",
                "operationId": "3ccc8c9c2d6659ffcafdbacc820e7ccf",
                "responses": {
                    "200": {
                        "description": "Daftar bimbingan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PenempatanMagang"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profil guru tidak ditemukan."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/guru/siswa/{id}/input-nilai": {
            "post": {
                "tags": [
                    "Guru"
                ],
                "summary": "Input nilai akhir & sahkan kelulusan siswa",
                "description": "Menyimpan nilai akhir (0-100), menandai penempatan lulus_magang, dan mengubah status siswa menjadi lulus.",
                "operationId": "25f07ccd99d3a9fb09616b9a2a8dd36a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID siswa.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nilai_akhir"
                                ],
                                "properties": {
                                    "nilai_akhir": {
                                        "type": "integer",
                                        "example": 90,
                                        "maximum": 100,
                                        "minimum": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nilai berhasil disimpan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PenempatanMagang"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Siswa tidak berada dalam bimbingan guru."
                    },
                    "422": {
                        "description": "Validasi gagal."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/siswa/absensi": {
            "get": {
                "tags": [
                    "Siswa"
                ],
                "summary": "Riwayat absensi siswa",
                "description": "Riwayat presensi milik siswa yang sedang login, urut tanggal terbaru.",
                "operationId": "fe88cbf5ee721f59fe1c6ed33fbd5102",
                "responses": {
                    "200": {
                        "description": "Riwayat absensi.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Absensi"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Siswa"
                ],
                "summary": "Catat presensi / izin",
                "description": "Absen masuk/pulang atau izin/sakit hari ini. `foto` dikirim sebagai multipart/form-data (opsional). Satu baris absensi per hari per siswa (di-update bila sudah ada).",
                "operationId": "ca8740da926262a463b66fb6f46fc72b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "type"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "masuk",
                                            "pulang",
                                            "izin_sakit"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "nullable": true,
                                        "enum": [
                                            "hadir",
                                            "sakit",
                                            "izin"
                                        ]
                                    },
                                    "catatan": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "foto": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Presensi berhasil dicatat.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Absensi"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/siswa/dashboard": {
            "get": {
                "tags": [
                    "Siswa"
                ],
                "summary": "Statistik dashboard siswa",
                "description": "Ringkasan status magang, penempatan, pengajuan terakhir, absensi hari ini, serta rekap presensi & jurnal siswa yang sedang login.",
                "operationId": "f27d01db57c51a8fd5069605be206c49",
                "responses": {
                    "200": {
                        "description": "Statistik dashboard siswa."
                    },
                    "404": {
                        "description": "Profil siswa tidak ditemukan."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PenempatanMagang": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "siswa_id": {
                        "type": "integer"
                    },
                    "tempat_magang_id": {
                        "type": "integer"
                    },
                    "guru_id": {
                        "type": "integer"
                    },
                    "tanggal_mulai": {
                        "type": "string",
                        "format": "date"
                    },
                    "tanggal_selesai": {
                        "type": "string",
                        "format": "date"
                    },
                    "nilai_akhir": {
                        "type": "integer",
                        "nullable": true
                    },
                    "status_pengesahan": {
                        "type": "string",
                        "enum": [
                            "belum_disahkan",
                            "disahkan",
                            "lulus_magang"
                        ]
                    }
                },
                "type": "object"
            },
...
            "Absensi": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "siswa_id": {
                        "type": "integer"
                    },
                    "tanggal": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-08-30"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "hadir",
                            "sakit",
                            "izin",
                            "alfa"
                        ]
                    },
                    "jam_masuk": {
                        "type": "string",
                        "nullable": true
                    },
                    "jam_pulang": {
                        "type": "string",
                        "nullable": true
                    },
                    "foto_masuk": {
                        "type": "string",
                        "nullable": true
                    },
                    "foto_pulang": {
                        "type": "string",
                        "nullable": true
                    },
                    "status_validasi_guru": {
                        "type": "string",
                        "enum": [
                            "menunggu",
                            "disetujui",
                            "ditolak"
                        ]
                    },
                    "catatan_guru": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Masukkan token dengan format: Bearer <token> (didapat dari endpoint POST /api/login).",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}